# docs/swagger.yaml
basePath: /
definitions:
  models.CreatePersonRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      surname:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.PersonResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      surname:
        type: string
    type: object
  models.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    type: object
  models.TokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Kişi yönetimi ve kimlik doğrulama için basit API.
  title: Go Kisi API
  version: "1.0"
paths:
  /add:
    post:
      consumes:
      - application/json
      description: JSON body ile yeni bir kişi ekler, email benzersiz olmalıdır
      parameters:
      - description: Kişi
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/models.CreatePersonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PersonResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Yeni kişi ekle / kayıt ol
      tags:
      - people
  /all:
    get:
      description: Veritabanındaki tüm kişileri döndürür
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PersonResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Tüm kişileri getir
      tags:
      - people
  /delete:
    get:
      description: ID'ye göre kişiyi siler
      parameters:
      - description: Kişi ID
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Kişi sil
      tags:
      - people
  /get:
    get:
      description: Verilen ID'ye göre kişiyi döndürür
      parameters:
      - description: Kişi ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PersonResponse'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: ID'ye göre kişi getir
      tags:
      - people
  /login:
    post:
      consumes:
      - application/json
      description: Email ve şifre ile giriş yapar, access ve refresh token döner
      parameters:
      - description: Giriş bilgileri
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Kullanıcı girişi
      tags:
      - auth
  /logout:
    post:
      consumes:
      - application/json
      description: Refresh token'ı revoke ederek kullanıcının çıkış yapmasını sağlar
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Çıkış yap
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Geçerli bir refresh token ile yeni access token döner
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Access token yenile
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"